requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cfiryptography==2.3
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
idna==2.7
pycparser==2.18
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
